+++ examples/get-started/tailwind.config.ts
@@ -2,7 +2,7 @@ import {sqlroomsTailwindPreset} from '@sqlrooms/ui';
 import type {Config} from 'tailwindcss';
 
 const config = {
-  presets: [sqlroomsTailwindPreset({})],
+  presets: [sqlroomsTailwindPreset()],
   content: [
     'src/**/*.{ts,tsx}',
     './node_modules/@sqlrooms/*/dist/**/*.js',