+++ packages/duckdb/src/DuckDbSlice.ts
@@ -1,8 +1,12 @@
-import {createBaseSlice, RoomState, useBaseRoomStore} from '@sqlrooms/core';
+import {DuckDbSliceConfig} from '@sqlrooms/duckdb-config';
+import {
+  createBaseSlice,
+  RoomState,
+  useBaseRoomStore,
+} from '@sqlrooms/room-store';
 import * as arrow from 'apache-arrow';
 import deepEquals from 'fast-deep-equal';
 import {produce} from 'immer';
-import {z} from 'zod';
 import {StateCreator} from 'zustand';
 import {DuckDbConnector, QueryHandle} from './connectors/DuckDbConnector';
 import {createWasmDuckDbConnector} from './connectors/createDuckDbConnector';
@@ -15,24 +19,13 @@ import {
   QualifiedTableName,
   splitSqlStatements,
 } from './duckdb-utils';
-import {createDbSchemaTrees as createDbSchemaTrees} from './schemaTree';
-import {DataTable, TableColumn, DbSchemaNode} from './types';
-
-export const DuckDbSliceConfig = z.object({
-  // nothing yet
-});
-export type DuckDbSliceConfig = z.infer<typeof DuckDbSliceConfig>;
+import {createDbSchemaTrees} from './schemaTree';
+import {DataTable, DbSchemaNode, TableColumn} from './types';
 
 export type SchemaAndDatabase = {
   schema?: string;
   database?: string;
 };
-
-export function createDefaultDuckDbConfig(): DuckDbSliceConfig {
-  return {
-    // nothing yet
-  };
-}
 
 /**
  * State and actions for the DuckDB slice
@@ -306,7 +299,7 @@ export function createDuckDbSlice({
               `CREATE OR REPLACE TABLE ${qualifiedName} AS (
               ${statements[0]}
             )`,
-            ).result,
+            ),
           );
           return {tableName, rowCount};
         },
@@ -349,7 +342,7 @@ export function createDuckDbSlice({
               database,
               table,
             })}`,
-          ).result;
+          );
           return getColValAsNumber(result);
         },
 
@@ -380,7 +373,7 @@ export function createDuckDbSlice({
                     .join(' AND ')}`
                 : ''
             }`,
-          ).result;
+          );
 
           const newTables: DataTable[] = [];
           for (let i = 0; i < describeResults.numRows; i++) {
@@ -427,8 +420,7 @@ export function createDuckDbSlice({
           const qualifiedTable = isQualifiedTableName(tableName)
             ? tableName
             : makeQualifiedTableName({table: tableName});
-          await connector.query(`DROP TABLE IF EXISTS ${qualifiedTable};`)
-            .result;
+          await connector.query(`DROP TABLE IF EXISTS ${qualifiedTable};`);
           await get().db.refreshTableSchemas();
         },
 
@@ -498,7 +490,7 @@ export function createDuckDbSlice({
             const connector = await get().db.getConnector();
             const result = await connector.query(
               `SELECT current_schema() AS schema, current_database() AS database`,
-            ).result;
+            );
             set((state) =>
               produce(state, (draft) => {
                 draft.db.currentSchema = result.getChild('schema')?.get(0);
@@ -533,7 +525,7 @@ export function createDuckDbSlice({
           const parsedQuery = (
             await connector.query(
               `SELECT json_serialize_sql(${escapeVal(sql)})`,
-            ).result
+            )
           )
             .getChildAt(0)
             ?.get(0);