+++ packages/jest-config/node.js
@@ -1,7 +1,0 @@
-import baseConfig from './base.js';
-
-/** @type {import('ts-jest').JestConfigWithTsJest} */
-export default {
-  ...baseConfig,
-  testEnvironment: 'node',
-};