+++ packages/room-config/src/BaseRoomConfig.ts
@@ -1,6 +1,6 @@
 import {z} from 'zod';
 import {DataSource} from './DataSource';
-import LayoutConfig, {DEFAULT_MOSAIC_LAYOUT} from './LayoutConfig';
+import {DEFAULT_MOSAIC_LAYOUT, LayoutConfig} from '@sqlrooms/layout-config';
 
 export const DEFAULT_ROOM_TITLE = 'Untitled room';