+++ packages/utils/jest.config.js
@@ -1,4 +1,4 @@
-import nodeConfig from '@sqlrooms/jest-config/node.js';
+import nodeConfig from '@sqlrooms/preset-jest/node.js';
 
 /** @type {import('ts-jest').JestConfigWithTsJest} */
 export default {